+++ chemics/main.py
@@ -244,8 +244,8 @@ class MainView(Qg.QMainWindow):  # REVIEW Code Class
         Exports the final kappa project data
         """
         # TODO issues/11
-        # noinspection PyCallByClass
-        file_name = self.controller.project_folder + "/kappa_" + self.controller.experiment_date.replace("/", ".") + ".csv"
+        file_name = self.controller.project_folder + "/kappa_"
+        file_name += self.controller.experiment_date.replace("/", ".") + ".csv"
         # noinspection PyCallByClass
         export_file = Qg.QFileDialog.getSaveFileName(self, "Save file", file_name, "*.csv")[0]
         if export_file: